Jeevanta Agent/requirements.txt
google-adk
python-dotenv
pymongo
apscheduler<4
fastapi
uvicorn[standard]
pyngrok
requests
